intermediate -> HeadstartRunner + cleanup
