--- vis/js/intermediate.js
@@ -1,316 +1,0 @@
-import ReactDOM from "react-dom";
-import React from "react";
-import { createStore } from "redux";
-import { Provider } from "react-redux";
-import $ from "jquery";
-
-import rootReducer from "./reducers";
-import {
-  initializeStore,
-  updateDimensions,
-  applyForceAreas,
-  applyForcePapers,
-  preinitializeStore,
-  zoomIn,
-  zoomOut,
-} from "./actions";
-
-import applyHeadstartMiddleware from "./middleware";
-
-import { applyForce } from "./utils/force";
-
-import { getChartSize, getListSize } from "./utils/dimensions";
-import Headstart from "./components/Headstart";
-import { createAnimationCallback } from "./utils/eventhandlers";
-import { removeQueryParams } from "./utils/url";
-import debounce from "./utils/debounce";
-import DataManager from "./dataprocessing/managers/DataManager";
-import FetcherFactory from "./dataprocessing/fetchers/FetcherFactory";
-
-// TODO use an npm package instead
-import { BrowserDetect } from "exports-loader?exports=BrowserDetect!../lib/browser_detect.js";
-
-/**
- * Class to sit between the "old" mediator and the
- * "new" modern frontend.
- *
- * This class should ideally only talk to the mediator and redux
- */
-class Intermediate {
-  constructor(config) {
-    this.config = config;
-    this.dataManager = new DataManager(config);
-
-    this.originalTitle = "";
-    this.actionQueue = [];
-
-    const middleware = applyHeadstartMiddleware(this);
-
-    this.store = createStore(rootReducer, middleware);
-  }
-
-  render() {
-    this.renderFrontend();
-    this.initStore();
-  }
-
-  renderFrontend() {
-    const container = $("#" + this.config.tag);
-    container.addClass("headstart");
-    container.append('<div id="app-container"></div>');
-
-    this.originalTitle = document.title;
-
-    this.store.dispatch(preinitializeStore(this.config));
-
-    ReactDOM.render(
-      <Provider store={this.store}>
-        <Headstart />
-      </Provider>,
-      document.getElementById("app-container")
-    );
-
-    window.addEventListener(
-      "popstate",
-      debounce(this.onBackButtonClick.bind(this), 300)
-    );
-
-    this.checkBrowserVersion();
-  }
-
-  // TODO split to multiple functions
-  async initStore() {
-    const config = this.config;
-
-    const dataFetcher = FetcherFactory.getInstance(config.mode, {
-      serverUrl: config.server_url,
-      files: config.files,
-      isStreamgraph: config.is_streamgraph,
-    });
-    const backendData = await dataFetcher.getData();
-
-    this.dispatchDataEvent(backendData);
-
-    const { size, width, height } = getChartSize(config);
-
-    this.dataManager.parseData(backendData, size);
-
-    const context = this.dataManager.context;
-
-    const list = getListSize(config, context, size);
-
-    this.store.dispatch(
-      initializeStore(
-        config,
-        context,
-        this.dataManager.papers,
-        this.dataManager.areas,
-        this.dataManager.streams,
-        size,
-        width,
-        height,
-        list.height,
-        this.dataManager.scalingFactors
-      )
-    );
-
-    if (!this.config.is_streamgraph) {
-      this.applyForceLayout();
-    }
-
-    // enable this for ability to share link to a zoomed bubble / paper
-    // if (queryParams.has("paper")) {
-    //   this.selectUrlPaper();
-    // } else {
-    //   this.zoomUrlArea();
-    // }
-    // remove the following lines if the previous line is uncommented
-    const queryParams = new URLSearchParams(window.location.search);
-    const paramsToRemove = [];
-    if (queryParams.has("area")) {
-      paramsToRemove.push("area");
-    }
-    if (queryParams.has("paper")) {
-      paramsToRemove.push("paper");
-    }
-    if (paramsToRemove.length > 0) {
-      removeQueryParams(...paramsToRemove);
-    }
-
-    // window resizing listener
-    window.addEventListener("resize", () => this.updateDimensions());
-  }
-
-  /**
-   * Function for when the browser back button is clicked.
-   *
-   * This function is the reason zoom-in and zoom-out actions are also queued
-   * in the queue middleware.
-   *
-   * For a better user experience, it should be debounced.
-   */
-  onBackButtonClick() {
-    const queryParams = new URLSearchParams(window.location.search);
-
-    if (!queryParams.has("area")) {
-      if (this.config.is_streamgraph && queryParams.has("paper")) {
-        removeQueryParams("paper");
-      }
-      this.store.dispatch(
-        zoomOut(createAnimationCallback(this.store.dispatch), true)
-      );
-
-      return;
-    }
-
-    // this can be optimized: if the area is the same as before, simply
-    // deselect the paper or select a different one
-
-    if (queryParams.has("area") && !queryParams.has("paper")) {
-      this.zoomUrlArea();
-
-      return;
-    }
-
-    if (queryParams.has("paper")) {
-      this.selectUrlPaper();
-    }
-  }
-
-  /**
-   * Selects the paper that's specified in the query params.
-   */
-  selectUrlPaper() {
-    if (this.config.is_streamgraph) {
-      // paper cannot be selected in streamgraph
-      removeQueryParams("area", "paper");
-      return;
-    }
-
-    const params = new URLSearchParams(window.location.search);
-
-    if (!params.has("paper")) {
-      return;
-    }
-
-    const zoomedPaper = params.get("paper");
-
-    const paper = this.dataManager.papers.find(
-      (p) => p.safe_id === zoomedPaper
-    );
-
-    if (!paper) {
-      return;
-    }
-
-    this.store.dispatch(
-      zoomIn(
-        { title: paper.area, uri: paper.area_uri },
-        createAnimationCallback(this.store.dispatch),
-        this.store.getState().zoom,
-        true,
-        paper
-      )
-    );
-  }
-
-  /**
-   * Zooms into an area that's specified in the query params.
-   */
-  zoomUrlArea() {
-    const params = new URLSearchParams(window.location.search);
-
-    if (!params.has("area")) {
-      return;
-    }
-
-    const zoomedArea = params.get("area");
-
-    if (this.config.is_streamgraph) {
-      // the instant zoom in doesn't work for streamgraph, because the data is not processed here yet (this.streamData)
-      // proper data processing refactoring is necessary
-      // this is a workaround that simply zooms out
-      removeQueryParams("area");
-      this.store.dispatch(
-        zoomOut(createAnimationCallback(this.store.dispatch), true)
-      );
-      return;
-    }
-
-    const area = this.dataManager.papers.find((a) => a.area_uri == zoomedArea);
-
-    if (!area) {
-      return;
-    }
-
-    this.store.dispatch(
-      zoomIn(
-        { title: area.area, uri: area.area_uri },
-        createAnimationCallback(this.store.dispatch),
-        this.store.getState().zoom,
-        true
-      )
-    );
-  }
-
-  // triggered on window resize
-  updateDimensions() {
-    const chart = getChartSize(this.config, this.dataManager.context);
-    const list = getListSize(this.config, this.dataManager.context, chart.size);
-    this.store.dispatch(updateDimensions(chart, list));
-  }
-
-  // used after each INITIALIZE
-  applyForceLayout() {
-    const state = this.store.getState();
-    applyForce(
-      state.areas.list,
-      state.data.list,
-      state.chart.width,
-      (newAreas) =>
-        this.store.dispatch(applyForceAreas(newAreas, state.chart.height)),
-      (newPapers) =>
-        this.store.dispatch(applyForcePapers(newPapers, state.chart.height)),
-      this.config
-    );
-  }
-
-  checkBrowserVersion() {
-    const browser = BrowserDetect.browser;
-    if (!["Chrome", "Firefox", "Safari"].includes(browser)) {
-      const alertMsg =
-        "You are using an unsupported browser. " +
-        "This visualization was successfully tested " +
-        "with the latest versions of Firefox, Chrome, Safari, " +
-        "Opera and Edge.";
-      alert(alertMsg);
-    }
-  }
-
-  /**
-   * Dispatches a custom event that the data has been loaded for reuse outside of Headstart.
-   * @param {Object} data the raw data
-   */
-  dispatchDataEvent(data) {
-    const elem = document.getElementById(this.config.tag);
-    const event = new CustomEvent("headstart.data.loaded", {
-      detail: { data },
-    });
-    elem.dispatchEvent(event);
-  }
-
-  rescaleMap(scaleBy, baseUnit, isContentBased, initialSort) {
-    this.config.scale_by = scaleBy;
-    this.config.base_unit = baseUnit;
-    this.config.content_based = isContentBased;
-    this.config.initial_sort = initialSort;
-    // TODO this might cause a bug (or there might be more params that require setting to false)
-    // bug description: map different after rescaling to the same metric
-    this.config.dynamic_sizing = false;
-
-    // TODO this can be optimized (omit another download)
-    this.initStore();
-  }
-}
-
-export default Intermediate;
